fix(channels): preserve dispatch contracts after flattening
