+++ src/plugin-sdk/channel-inbound.ts
@@ -1,4 +1,4 @@
-import type { DispatchFromConfigResult } from "../auto-reply/reply/dispatch-from-config.types.js";
+import type { DispatchFromConfigResult } from "../auto-reply/reply/dispatch-from-config.js";
 // Channel inbound contracts define plugin ingress payloads and reply dispatch metadata.
 import {
   buildChannelInboundEventContext,
